requirements.txt
openai==1.34.0
python-dotenv==1.0.0
langchain == 0.2.5
langchain-community == 0.2.5
langchain-core == 0.2.7
langchain-openai == 0.1.8
langchain-text-splitters == 0.2.1
streamlit
azure-storage-blob
azure-identity
unstructured
